feat: Improve loading UI and adjust API endpoint
