+++ src/app/(affiliate)/affiliate/page.tsx
@@ -16,10 +16,10 @@ export default function DashboardPage() {
             <div className="flex flex-col gap-4 p-4 pt-0">
               <div className="min-h-[100vh] flex-1 rounded-xl bg-muted/50 md:min-h-min">
                 <div className="p-6">
-                  <div className="mb-8">
-                    <h1 className="text-3xl font-bold tracking-tight">Affiliate Dashboard</h1>
+                  <div className="mb-6">
+                    <h1 className="text-3xl font-bold tracking-tight">Performance Analytics</h1>
                     <p className="text-muted-foreground">
-                      Track your affiliate performance
+                      Detailed metrics for your affiliate link performance and revenue tracking
                     </p>
                   </div>
                   <PerformanceAnalytics />
